Move bundlesize plugin into separate file
